updated page for add and edit network
