+++ ui/pages/settings/networks-tab/index.scss
@@ -86,6 +86,10 @@
       position: sticky;
       bottom: 0;
       box-shadow: 0 0 8px 0 var(--color-shadow-default);
+
+      &--page {
+        box-shadow: none;
+      }
     }
   }
